# app/_archive/esgss_junaikey_beta/openapi.yaml
openapi: 3.0.3
info:
  title: Omni Celestial Server API - v10.0-omni
  description: |-
    API for the ESGss JunAiKey Beta (Omni Server). 
    Provides endpoints for managing ESG metrics, AI agents, knowledge base, swarm intelligence, and more, in compliance with the Omni Agent Protocol (5T).
  version: 10.0.0-omni
servers:
  - url: https://esg-dashboard-c3ytffo5qq-de.a.run.app
    description: Production Server
  - url: http://localhost:3001
    description: Development Server

# 1. Define Security Scheme
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT # or Opaque, depending on the token type
  schemas:
...
    KnowledgeIngestResponse:
      type: object
      properties:
        status:
          type: string
          example: 'success'
        message:
          type: string
          example: 'Knowledge ingested successfully'
        chunkId:
          type: string
          example: 'vector_db_chunk_id'
        source:
          type: string
          example: 'https://example.com/report.pdf'
...
    SwarmResponse:
      type: object
      properties:
        status:
          type: string
          example: 'success'
        finalOutput:
          type: string
          example: 'The final summary report...'
        plan:
          type: string
          example: 'Step 1: Research regulations. Step 2: ...'
        history:
          type: array
          items:
            type: object
...
security:
  - bearerAuth: []
paths:
...
  /api/manifest:
    post:
      summary: Manifest AI Agent
      description: Creates a new Agent session and "awakens" an AI model instance for it.
      tags:
        - Core API
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                source_agent:
                  oneOf:
                    - type: string
                    - type: object
                      properties:
                        name:
                          type: string
                        system_prompt:
                          type: string
                overrides:
                  type: object
                  properties:
                    mask:
                      type: object
                      properties:
                        tone:
                          type: string
                        language:
                          type: string
      responses:
        '200':
          description: Agent session created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                  status:
                    type: string
                  agentName:
                    type: string
                  agentId:
                    type: string
...
  /api/interact:
    get:
      summary: Interact with an Agent
      description: Engages in a real-time interaction with an awakened Agent session using Server-Sent Events (SSE).
      tags:
        - Core API
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
        - name: message
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: An SSE stream of interaction events.
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  data: {"type": "thought", "content": "Thinking..."}
                  data: {"type": "text", "content": "Response part 1"}
  /api/learn:
    post:
      summary: Ingest New Knowledge
      description: Injects new knowledge into a specified Knowledge Base.
      tags:
        - Core API
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                kbId:
                  type: string
                text:
                  type: string
                source:
                  type: string
                metadata:
                  type: object
      responses:
        '200':
          description: Knowledge ingested successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeIngestResponse'
...
  /api/swarm:
    post:
      summary: Launch Swarm Intelligence
      description: Initiates a LangGraph-based agent swarm to solve a complex goal.
      tags:
        - Advanced API
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                goal:
                  type: string
      responses:
        '200':
          description: Swarm task completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwarmResponse'
...